move test_gatedgcnconv
